requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
idna==3.10
mercantile==1.2.1
numpy==2.2.4
pillow==11.1.0
pyparsing==3.2.3
pyproj==3.7.1
rasterio==1.4.3
requests==2.32.3
scipy==1.15.2
urllib3==2.3.0
